move to struct config file
